Return unformatted text and omit error output on exception
When using yapf as an autoformatter in vim, syntax errors cause the
buffer to be overwritten with the backtrace. While it's nice to see
a backtrace without having to run the code, it's pretty annoying to
not be able to save python files with syntax errors.

This changeset forces output of the original, unformatted code on
any uncaught exceptions, omitting the backtrace entirely.

Below is how I use yapf in my vim config.

```
" ~/.vim/filetype.vim
autocmd BufWritePre * call OnBufWritePre()
function! OnBufWritePre()
    if &filetype=='python'
        let view = winsaveview()
        %!yapf
        call winrestview(view)
    endif
endfunction
```